requirements.txt
pandas>=1.5.0
matplotlib>=3.6.0
seaborn>=0.12.0
numpy>=1.24.0
openpyxl>=3.0.0
streamlit>=1.28.0
plotly>=5.15.0
flask>=2.3.0
mysql-connector-python>=8.0.0
sqlalchemy>=1.4.0 